+++ packages/substrate-bindings/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./ss58-util"
+export * from "./multisig"